Expose openFileWithCloseOnExec and openExistingFileWithCloseOnExec
Fixes #21
