+++ file-io.cabal
@@ -1,6 +1,6 @@
 cabal-version:      2.4
 name:               file-io
-version:            0.1.1
+version:            0.1.2
 synopsis:           Basic file IO operations via 'OsPath'
 description:        Basic file IO operations like Prelude, but for 'OsPath'.
 homepage:           https://github.com/hasufell/file-io
@@ -53,7 +53,9 @@ library
 
   exposed-modules:
     System.File.OsPath
+    System.File.OsPath.Internal
     System.File.PlatformPath
+    System.File.PlatformPath.Internal
 
   other-modules: System.File.Platform
